requirements.txt
beautifulsoup4
discord.py
facebook-sdk
instaloader
langdetect
linkedin-api
matplotlib
nltk
numpy
pandas
praw
python-dotenv
requests
schedule
scikit-learn
snscrape
sqlalchemy
telethon
textblob
tweepy
TikTokApi
